Refactoring-AbstractEmployment
