+++ module/Application/language/xx_XX.po
@@ -268,6 +268,9 @@ msgstr "cv-mainHeader-personalData-years"
 msgid "cv-careerGoals-sectionTitle"
 msgstr "cv-careerGoals-sectionTitle"
 
+msgid "cv-careerGoals-recipent"
+msgstr "cv-careerGoals-recipent"
+
 msgid "cv-careerGoals-content"
 msgstr "cv-careerGoals-content"
 
@@ -295,20 +298,17 @@ msgstr "cv-hobby-sectionTitle"
 msgid "cv-hobby-content"
 msgstr "cv-hobby-content"
 
-msgid ""
-msgstr ""
+msgid "cv-softwareTools-sectionTitle"
+msgstr "cv-softwareTools-sectionTitle"
 
-msgid ""
-msgstr ""
+msgid "cv-languages-sectionTitle"
+msgstr "cv-languages-sectionTitle"
 
-msgid ""
-msgstr ""
+msgid "cv-skills-shortcut-months"
+msgstr "cv-skills-shortcut-months"
 
-msgid ""
-msgstr ""
-
-msgid ""
-msgstr ""
+msgid "cv-skills-shortcut-years"
+msgstr "cv-skills-shortcut-years"
 
 msgid ""
 msgstr ""
