+++ module/Application/src/Element/MainHeaderMostRecentInfo.php
@@ -24,11 +24,25 @@ class MainHeaderMostRecentInfo extends AbstractTcpdfDecorator
      */
     public function renderMostRecentInfo()
     {
+        $this->configure();
+
         $this->tcpdf->SetXY(
             self::INFO_CURSOR_X,
             self::INFO_CURSOR_Y
         );
 
+        $this->tcpdf->Write(
+            self::INFO_FONT_SIZE,
+            $this->trans('cv-mainHeader-mostRecentVersion'),
+            PdfConfig::DOCUMENT_URL
+        );
+    }
+
+    /**
+     * Configures element
+     */
+    private function configure()
+    {
         $this->tcpdf->SetTextColor(
             Color::TEXT_COLOR_LIGHT_RED,
             Color::TEXT_COLOR_LIGHT_GREEN,
@@ -39,12 +53,6 @@ class MainHeaderMostRecentInfo extends AbstractTcpdfDecorator
             $this->tcpdf->tahoma,
             Font::BOLD,
             self::INFO_FONT_SIZE
-        );
-
-        $this->tcpdf->Write(
-            self::INFO_FONT_SIZE,
-            $this->trans('cv-mainHeader-mostRecentVersion'),
-            PdfConfig::DOCUMENT_URL
         );
     }
 }
